requirements.txt
python-decouple==3.8
SQLAlchemy==2.0.31
pydantic==2.8.2
aiosqlite==0.20.0
APScheduler==3.10.4
uvicorn==0.30.6
fastapi-responses==0.2.1
fastapi==0.115.0
python-jose==3.3.0